requirements.txt
Flask==0.12.2
pycryptodome